refactor(OptionsConfigurationComponent): use new ConfigurationTableComponent
